+++ packages/core/src/common/point-nd-util.ts
